# sales-service/target/classes/documentacion/swagger25-08.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/productos/update/{id}": {
            "put": {
                "tags": [
                    "producto-controller"
                ],
                "operationId": "actualizarProducto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductoDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos/compra/{id}": {
            "put": {
                "tags": [
                    "producto-controller"
                ],
                "operationId": "compraProducto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "cantidad",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos/add": {
            "post": {
                "tags": [
                    "producto-controller"
                ],
                "operationId": "insertarProducto",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductoDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos/{id}": {
            "get": {
                "tags": [
                    "producto-controller"
                ],
                "operationId": "obtenerProductoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductoDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "producto-controller"
                ],
                "operationId": "deleteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/productos/all": {
            "get": {
                "tags": [
                    "producto-controller"
                ],
                "operationId": "obtenerTodos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductoDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos/": {
            "get": {
                "tags": [
                    "producto-controller"
                ],
                "operationId": "home",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ProductoDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nombre": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            }
        }
    }
}